unified_agent/requirements.txt
# Unified AI Agent Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
cryptography>=42.0.0

# AI/LLM
openai>=1.10.0
pinecone-client>=3.0.0

# Telephony
twilio>=8.10.0

# HTTP Client
httpx>=0.26.0

# SSH Tunneling
sshtunnel>=0.4.0

# Authentication
pyjwt>=2.8.0
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Background Tasks
celery>=5.3.0
flower>=2.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
